Chnages made
I realise that alot of this is my coding and it may not be optimal or even clear. I added some comments to help with this and tried to make the code a little more readable.
If anyone needs clarification let me know. Also freely modify or completely change to be optimal/to fit your needs or standard.

Chnages
-Added Map Fragment functionality(random locations based on current location)

-Dummy Scanning Activity + xml added

-Comments added

-Increased readability of code