Seeing, not believeing
Was able to get the list of players to show for the score ranking.

It is the wrong output however. I am struggling to query the db properly to get the top 10. The worst part is, it was working perfectly at one point. Then all of a sudden nothing makes sense.

I will take another look later. Might be an accident delete of a line of code or maybe I can get someone else to do the querying.